Refactor LevelOrchestrator to use Facade Pattern
Abstract responsibilities into separate classes
* LevelGenerationManager
* LevelTransitionManager
